Joysticks: Move interfaces to interfaces/ subfolder
