+++ xbmc/games/controllers/dialogs/GUIDialogButtonCapture.cpp
@@ -21,9 +21,9 @@
 #include "GUIDialogButtonCapture.h"
 #include "guilib/GUIWindowManager.h"
 #include "guilib/WindowIDs.h"
+#include "input/joysticks/interfaces/IButtonMap.h"
+#include "input/joysticks/interfaces/IButtonMapCallback.h"
 #include "input/joysticks/JoystickIDs.h"
-#include "input/joysticks/IButtonMap.h"
-#include "input/joysticks/IButtonMapCallback.h"
 #include "input/joysticks/JoystickUtils.h"
 #include "input/IKeymap.h"
 #include "input/ActionIDs.h"
