+++ xbmc/games/controllers/dialogs/GUIDialogButtonCapture.h
@@ -19,7 +19,7 @@
  */
 #pragma once
 
-#include "input/joysticks/IButtonMapper.h"
+#include "input/joysticks/interfaces/IButtonMapper.h"
 #include "threads/Event.h"
 #include "threads/Thread.h"
 #include "utils/Observer.h"
