+++ xbmc/games/controllers/dialogs/GUIDialogIgnoreInput.cpp
@@ -20,8 +20,8 @@
 
 #include "GUIDialogIgnoreInput.h"
 #include "guilib/LocalizeStrings.h"
-#include "input/joysticks/IButtonMap.h"
-#include "input/joysticks/IButtonMapCallback.h"
+#include "input/joysticks/interfaces/IButtonMap.h"
+#include "input/joysticks/interfaces/IButtonMapCallback.h"
 #include "input/joysticks/JoystickTranslator.h"
 #include "utils/log.h"
 #include "utils/StringUtils.h"
