+++ xbmc/games/controllers/windows/GUIConfigurationWizard.h
@@ -21,7 +21,7 @@
 
 #include "IConfigurationWindow.h"
 #include "input/joysticks/DriverPrimitive.h"
-#include "input/joysticks/IButtonMapper.h"
+#include "input/joysticks/interfaces/IButtonMapper.h"
 #include "input/keyboard/IKeyboardHandler.h"
 #include "input/mouse/IMouseInputHandler.h"
 #include "threads/CriticalSection.h"
