+++ xbmc/input/joysticks/CMakeLists.txt
@@ -6,19 +6,19 @@ set(SOURCES DeadzoneFilter.cpp
             JoystickUtils.cpp
             RumbleGenerator.cpp)
 
-set(HEADERS interfaces/IKeyHandler.h
+set(HEADERS interfaces/IButtonMap.h
+            interfaces/IButtonMapCallback.h
+            interfaces/IButtonMapper.h
+            interfaces/IButtonSequence.h
+            interfaces/IDriverHandler.h
+            interfaces/IDriverReceiver.h
+            interfaces/IInputHandler.h
+            interfaces/IInputProvider.h
+            interfaces/IInputReceiver.h
+            interfaces/IKeyHandler.h
             interfaces/IKeymapHandler.h
             DeadzoneFilter.h
             DriverPrimitive.h
-            IButtonMap.h
-            IButtonMapCallback.h
-            IButtonMapper.h
-            IButtonSequence.h
-            IDriverHandler.h
-            IDriverReceiver.h
-            IInputHandler.h
-            IInputProvider.h
-            IInputReceiver.h
             JoystickEasterEgg.h
             JoystickIDs.h
             JoystickMonitor.h
