+++ xbmc/input/joysticks/DeadzoneFilter.cpp
@@ -20,7 +20,7 @@
 
 #include "DeadzoneFilter.h"
 #include "JoystickIDs.h"
-#include "IButtonMap.h"
+#include "input/joysticks/interfaces/IButtonMap.h"
 #include "peripherals/devices/Peripheral.h"
 #include "utils/log.h"
