+++ xbmc/input/joysticks/RumbleGenerator.cpp
@@ -21,7 +21,7 @@
 #include "RumbleGenerator.h"
 #include "games/controllers/Controller.h"
 #include "games/controllers/ControllerManager.h"
-#include "input/joysticks/IInputReceiver.h"
+#include "input/joysticks/interfaces/IInputReceiver.h"
 #include "input/joysticks/JoystickIDs.h"
 #include "ServiceBroker.h"
