+++ xbmc/input/joysticks/generic/InputHandling.cpp
@@ -20,8 +20,8 @@
 
 #include "InputHandling.h"
 #include "input/joysticks/dialogs/GUIDialogNewJoystick.h"
+#include "input/joysticks/interfaces/IButtonMap.h"
 #include "input/joysticks/DriverPrimitive.h"
-#include "input/joysticks/IButtonMap.h"
 #include "input/joysticks/JoystickUtils.h"
 #include "utils/log.h"
