+++ xbmc/input/joysticks/generic/InputHandling.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "FeatureHandling.h"
-#include "input/joysticks/IDriverHandler.h"
+#include "input/joysticks/interfaces/IDriverHandler.h"
 #include "input/joysticks/JoystickTypes.h"
 
 #include <map>
