+++ xbmc/input/joysticks/interfaces/IButtonMapCallback.h
