+++ xbmc/input/joysticks/interfaces/IButtonMapper.h
