+++ xbmc/input/joysticks/interfaces/IInputProvider.h
