+++ xbmc/peripherals/addons/AddonButtonMapping.cpp
@@ -20,7 +20,7 @@
 
 #include "AddonButtonMapping.h"
 #include "input/joysticks/generic/ButtonMapping.h"
-#include "input/joysticks/IButtonMapper.h"
+#include "input/joysticks/interfaces/IButtonMapper.h"
 #include "peripherals/addons/AddonButtonMap.h"
 #include "peripherals/Peripherals.h"
 #include "utils/log.h"
