+++ xbmc/peripherals/addons/AddonInputHandling.cpp
@@ -21,8 +21,8 @@
 #include "AddonInputHandling.h"
 #include "input/joysticks/generic/DriverReceiving.h"
 #include "input/joysticks/generic/InputHandling.h"
-#include "input/joysticks/IInputHandler.h"
-#include "input/joysticks/IDriverReceiver.h"
+#include "input/joysticks/interfaces/IInputHandler.h"
+#include "input/joysticks/interfaces/IDriverReceiver.h"
 #include "peripherals/addons/AddonButtonMap.h"
 #include "peripherals/devices/PeripheralJoystick.h"
 #include "peripherals/Peripherals.h"
