+++ xbmc/peripherals/devices/PeripheralJoystick.cpp
@@ -19,9 +19,9 @@
  */
 
 #include "PeripheralJoystick.h"
+#include "input/joysticks/interfaces/IDriverHandler.h"
 #include "input/joysticks/keymaps/KeymapHandling.h"
 #include "input/joysticks/DeadzoneFilter.h"
-#include "input/joysticks/IDriverHandler.h"
 #include "input/joysticks/JoystickIDs.h"
 #include "input/joysticks/JoystickMonitor.h"
 #include "input/joysticks/JoystickTranslator.h"
