+++ xbmc/peripherals/devices/PeripheralJoystick.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "Peripheral.h"
-#include "input/joysticks/IDriverReceiver.h"
+#include "input/joysticks/interfaces/IDriverReceiver.h"
 #include "input/joysticks/JoystickTypes.h"
 #include "threads/CriticalSection.h"
